Update patch set 4
Patch Set 4: I would prefer that you didn't submit this

Presently fails the first time tethering is enabled, but works on subsequent attempts.

The issue is that the phone-side usb0 interface needs to be both up and have a ready link (rndis interface exposed and USB connected) for Tethering to recognize usb0 as a tetherable interface, and the link is not yet ready when Tethering "checks" the first time.

I'll look to see if there's a reasonable framework change that would avoid that problem.

Patch-set: 4