Update patch set 4
Patch Set 4:

A quick grep of the gingerbread repo says that we (non-Google Samsung devices) are more or less the only one using these specific CFLAGS.

The build system sets some CFLAGS in build/core/combo/arch/arm/armv7-a-neon.mk (from our TARGET_ARCH_VARIANT) and also sets ARCH_ARM_HAVE_ARMV7A, so we do not need to set it manually.

The only effective CFLAGS we're setting is -mtune=cortex-a8 and I don't know what effect that has, if it will have any effects at all. Google chose not to use it for crespo (or other armv7-a-neon), which is may be done intentionally or due to negligence. But a quick diff of two different builds, one using this flag and one not using it, says that the binaries have equal sizes (but slightly different contents).

So if a CFLAGS is worth keeping, it's -mtune=cortex-a8, but if we want to follow crespo, which has very similar hardware, we would simply remove it.

As for HGL, it's probably a very old left-over setting with absolutely no effects on our devices.

Patch-set: 4