Update patch set 1
Patch Set 1:

I thought epicmtd had an open source audio HAL just like the other aries devices?

This shouldn't be necessary, it's primarily a workaround for a bug with Samsung's blobs, and is not seen in opensource HALs.

The only devices I'm aware of that need this are Exynos4210 devices with Yamaha chipsets and maybe some of the Qcom devices with blob audio.

Patch-set: 1