Update patch set 2
Patch Set 2: Code-Review+1

Now it looks fine :)

To see that the orientation awareness works fine, try the DPAD keys in an app that enforces portrait - e.g. some views in Apollo. The function of DPAD keys will be changing with the changes of display orientation enforced by the app.

Patch-set: 2
Label: Code-Review=+1